Add site-actions login/logout links. Changed in two different places: Plone actions and content_render.xsl
But wait! Shouldn't you be changing that in three places, you ask? Indeed, that used to be the case, but I've set content_template to use the same macro as the rest of the Plone templates for site actions (just like I did for global tabs earlier.) Like in the global tabs case, I had to do a little manipulation of the content styles to match the Plone macro output, but since we only ask for 'first' (nee 'cnx_first') in context of another cnx_named element, I think it doesn't hurt our classname-independance at all.

Requires re-run of upActions.zctl.