Use one template (well, that calls in other templates) to generate the mailto: links for e-mailing the authors of content.
If they're e-mailing the author(s) of the module outside a collection 
context, the e-mail looks like:

  To: janedoe@example.com
  Cc: cnx@cnx.org
  Subject: Feedback on Connexions module: "Grilling steaks"
  Body:
    Feedback on Connexions module: "Grilling steaks"
    (http://cnx.org/content/mSTEAK/latest/)

If they're e-mailing the author(s) of the module in a collection context,
the e-mail looks like:

  To: janedoe@example.com
  Cc: cnx@cnx.org
  Subject: Feedback on Connexions module: "Grilling steaks", as used
    in collection: "All about steaks"
  Body:
    Feedback on Connexions module: "Grilling Steaks"
    (http://cnx.org/content/mSTEAK/latest/), as used in collection:
    "All about steaks" (http://cnx.org/content/colSTEAK/1.1)

If they're e-mailing the author(s) of the collection from a module, the
e-mail looks like:

  To: johndoe@example.com
  Cc: cnx@cnx.org
  Subject: Feedback on Connexions collection: "Grilling steaks",
    module: "All about steaks"
  Body:
    Feedback on Connexions collection: "All about steaks"
    (http://cnx.org/content/colSTEAK/1.1/), module: "Grilling steaks"
    (http://cnx.org/content/mSTEAK/latest/)

If they're e-mailing the author(s) of either the collection or the
module (the author(s) being the same set (in which case we only
provide a single link)), the e-mail looks like:

  To: janedoe@example.com
  Cc: cnx@cnx.org
  Subject: Feedback on Connexions module: "Grilling steaks", as used
    in collection: "All about steaks"
  Body:
    Feedback on Connexions module: "Grilling Steaks"
    (http://cnx.org/content/mSTEAK/latest/), as used in collection:
    "All about steaks" (http://cnx.org/content/colSTEAK/1.1)

i.e. it's pretty much the same as the 2nd one (the only difference
being that in this case janedoe@example.com also happened to author
"All about steaks").
 
Fixes #6254