Number and name exercises that are children of examples differently than exercises that are elsewhere in a document.  Fixes #6275
The new way layout of how empty cnxns are produced should make for an
easier merge w/ the stuff that was in cnxml for print-on-demand-1-1

While I'm here, I noticed that many CNXML elements don't have
translations of them if you were to make an empty to cnxn to them.
Most are unlikely cases, but might as well do the complete set.
Captialize the ones that are just putting through custom note and rule
types.