These changes make anchors for 0.5 <cnxn>s, 0.5 <link>s, and 0.6 <link>s, and expands the link text if the <cnxn>/<link> is empty.  The basic algorithm in most cases is "element name (unless there's a label override) followed by element number (w/ consideration for @type)". Nowhere near perfect, but a better algorithm can be worked out after the milestone.
This also changes <subfigure>s to be labeled "Figure 1(a)" instead of
"Subfigure 1.1".  Refs #5779.

This also should fix #6176 which is the problem w/ @version not being
honored in <cnxn> URLs.

This does not yet fix alt generation text (for new subfigure
numbering), or add linking functionality to elements like <cite>, 
<foreign>, <term>, and <quote>.