My first cut at handling some of the new media elements.  Doesn't yet handle <object>, <audio>, <flash>, <labview>, <download>, or <text>.
Fixed some logic in unibrowser for subfigure labeling and number
generation for empty <cnxn>/<link>s.  

Had to change the automatic alt generation (for images or applets that
don't have an alt param/attribute written by the user) because we're
no longer numbering them "Subfigure 1.1", but "Figure 1(a)".  
Refs #5779 but still can't call it a fix, because it's not very clear
what to do for this text and for empty cnxn expansion text when the 
<figure> and/or the <subfigure> or have a <label> and/or a @type.