Merging in MDML changes:   Merged revisions 24464-24804,24807-24863,24865-24928,24930-25079,25081-25100,25102-25105 via svnmerge from   https://software.cnx.rice.edu/svn/cnxml/branches/cnxml-0.6-mdml-0.5
Note that we made some changes in MDML over against what we merged from, but in order to commit a more or less working piece all at once, they are submerged in the general merge.

Also work on upconversion: handle new metadata temporarily by adding in dummy metadata elements for fields not now present in the module metadata; we will use a zpt to expose the needed metadata to the upconversion to get the real deal.  Unfortunately, that will make a complete upconversion of 7200+ modules (and hence testing) much slower.